src/requirements.txt
# Основные зависимости здесь
# python-telegram-bot
aiogram 
fastapi
uvicorn
# psycopg2-binary
# SQLAlchemy==2.0.25
alembic
# celery
# redis
loguru
python-telegram-bot==20.8
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.0
alembic==1.12.1
asyncpg==0.29.0
ruff==0.1.11
pre-commit==4.2.0
pydantic-settings
python-dotenv